requirements.txt
openai-whisper
torch
pyttsx3
sounddevice
scipy
Pillow
python-dotenv
fastapi
uvicorn
openai==1.30.1
jinja2

